Refactoring of perceptual metrics
